moved smallcontrol_fragment to listplayer
