+++ app/src/main/java/de/codingforcelm/idmp/fragment/ListPlayerFragment.java
@@ -4,6 +4,7 @@ import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ImageView;
 
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentTransaction;
@@ -16,6 +17,10 @@ import de.codingforcelm.idmp.player.service.MusicService;
 
 public class ListPlayerFragment extends Fragment {
     private List<PhysicalSong> songList;
+    private ImageView playPauseButton;
+    private ImageView nextButton;
+    private ImageView prevButton;
+    private MusicService service;
 
     public ListPlayerFragment(List<PhysicalSong> songList) {
         this.songList=songList;
@@ -32,17 +37,60 @@ public class ListPlayerFragment extends Fragment {
     @Override
     public void onViewCreated(View view, Bundle savedInstanceState) {
         FragmentTransaction ft = getChildFragmentManager().beginTransaction();
-        ft.replace(R.id.player_list_bot, new ControlsSmallFragment(), "CONTROLS_SMALL");
         ft.replace(R.id.player_list_top, new SongListFragment(songList), "SONGLIST");
         ft.commit();
+
+        playPauseButton = view.findViewById(R.id.lp_playPauseButton);
+        playPauseButton.setOnClickListener(new ListPlayerFragment.PlayPauseOnClickListener());
+        nextButton = view.findViewById(R.id.lp_nextButton);
+        nextButton.setOnClickListener(new ListPlayerFragment.NextOnClickListener());
+        prevButton = view.findViewById(R.id.lp_prevButton);
+        prevButton.setOnClickListener(new ListPlayerFragment.PrevOnClickListener());
     }
+
 
     /**
      * TODO doc
      * @param service service
      */
     public void setService(MusicService service) {
-        ControlsSmallFragment frag = (ControlsSmallFragment)getChildFragmentManager().findFragmentByTag("CONTROLS_SMALL");
-        frag.setService(service);
+        this.service=service;
+    }
+
+
+    private class NextOnClickListener implements View.OnClickListener {
+        @Override
+        public void onClick(View v) {
+            if(service != null){
+                service.nextSong();
+                playPauseButton.setImageResource(android.R.drawable.ic_media_pause);
+            }
+        }
+    }
+
+    private class PlayPauseOnClickListener implements View.OnClickListener {
+
+        @Override
+        public void onClick(View v) {
+            if (service != null) {
+                if (service.isPlaying()) {
+                    service.pauseSong(true);
+                    playPauseButton.setImageResource(android.R.drawable.ic_media_play);
+                } else {
+                    service.resumeSong();
+                    playPauseButton.setImageResource(android.R.drawable.ic_media_pause);
+                }
+            }
+        }
+    }
+
+    private class PrevOnClickListener implements View.OnClickListener {
+        @Override
+        public void onClick(View v) {
+            if(service != null){
+                service.prevSong();
+                playPauseButton.setImageResource(android.R.drawable.ic_media_pause);
+            }
+        }
     }
 }
