--- app/src/main/res/layout/fragment_controls_small.xml
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent" android:layout_height="match_parent"
-    android:orientation="vertical" xmlns:app="http://schemas.android.com/apk/res-auto">
-
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_weight="40"
-        android:orientation="horizontal">
-
-        <ImageView
-            android:id="@+id/lp_image"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_weight="2"
-            app:srcCompat="@android:drawable/ic_menu_report_image" />
-
-        <ImageView
-            android:id="@+id/lp_prevButton"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_weight="4"
-            app:srcCompat="@android:drawable/ic_media_previous" />
-
-        <ImageView
-            android:id="@+id/lp_playPauseButton"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_weight="4"
-            app:srcCompat="@android:drawable/ic_media_play" />
-
-        <ImageView
-            android:id="@+id/lp_nextButton"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_weight="4"
-            app:srcCompat="@android:drawable/ic_media_next" />
-
-    </LinearLayout>
-
-</LinearLayout>
